Update Parse.hs
It is convenient for parsing comlicated xml-formats to be able to skip tags that are not used. For example I need to skip <metadata>-tags in SVG because I don't use it (see diagrams-input). Not having this tagSkip function would mean having to write parsers for RDF inside of <metadata>