Consider whitespace when tokenising
This change improves tokenisation by handling whitespace as a semantic
feature. Instead of relying on the `tag` combinator to skip leading
whitespace, this is done during tokenisation.

A significant change here is that we need to now take context into
account when parsing a sequence of text content, since the leading node
may skip leading whitespace, whereas subsequent ones need to preserve
it.  This is handled by changing the type of parseToken to return a list
of tokens, so that consecutive text nodes can be all handled together,
with a different setting of the preserve-leading-whitespace flag on the
tail versus the head. The downside of this is that the event positions
will be less accurate (it is the aim of future work to improve this).

Another area for further work is the fact that for `<node> xyz </node>`
we will strip the leading space but keep the trailing one (yielding
"xyz "). This does not break any tests (the aim of this change is simply to
introduce this change while preserving API compatibility) but should in
the future be handled consistently, yielding "xyz".