Allow omission of the XML declaration when rendering documents.
In-band encoding information can be unreliable, so it can sometimes be
preferable to omit it entirely; content-sniffing for XML is similarly
unreliable. Without these considerations, there is no need for an XML
declaration. The default is to retain the XML declaration for bytewise
output stability.

In addition, the internal-only constructor TokenBeginDocument that
stores the XML declaration's attributes has been renamed
TokenXMLDeclaration to reduce confusion, as not every document begins
with an XML declaration; this was already the case for parsing, and is
now also the case for rendering.