Don't allow `[` or `]` in XML names.
This is an example of a DOCTYPE that was not being parsed correctly
before:

```
<!DOCTYPE language[
  <!ENTITY nmtoken "[\-\w\d\.:_]+">
  <!ENTITY entref  "(#[0-9]+|#[xX][0-9A-Fa-f]+|&nmtoken;);">
]>
```

xml-conduit was parsing `language[` as the root element name.

I have kept to the most minimal possible change in this
PR, because I don't want to break anything inadvertently. However,
the current parser is still far from correct. As I understand
it, only a few symbols (`_`, `-`, `.`) are allowed in element names
(in addition, `:` can be used for a namespace, but that is supported
separately in this parser).  The current parser would accept things
like `<foo~bar>`.